+++ ICD.Connect.Lighting.CrestronPro/EiscLightingAdapter/Components/IEiscLightingComponent.cs
@@ -1,0 +1,12 @@
+﻿using ICD.Connect.API.Nodes;
+
+namespace ICD.Connect.Lighting.CrestronPro.EiscLightingAdapter.Components
+{
+	public interface IEiscLightingComponent: IConsoleNode
+	{
+		int Id { get; }
+		string Name { get; }
+
+		LightingProcessorControl ToLightingProcessorControl();
+	}
+}